requirements.txt
coverage
django
djangorestframework
djoser
djangorestframework_simplejwt
django-environ
drf-yasg
black